[#30] Rename: write 관련 컴포넌트 write 폴더로 이동
